增加地图标牌的删除与新增的权限（对施工单位）

+++ src/main/java/com/sju/program/controller/SignController.java
@@ -1,6 +1,7 @@
 package com.sju.program.controller;
 
 import com.sju.program.domain.Project;
+import com.sju.program.domain.model.BaseUser;
 import com.sju.program.domain.model.LoginUser;
 import com.sju.program.domain.po.SignInfoPo;
 import com.sju.program.domain.sign.Sign;
@@ -83,6 +84,11 @@ public class SignController extends BaseController{
 	@ApiOperation(value = "增加地图标牌")
 	@PostMapping("/inMap")
 	public AjaxResult insertSignInMap(@RequestBody SignInfoPo signInfoPo){
+		LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+		BaseUser baseUser=(BaseUser)loginUser.getUser();
+		if (baseUser.getAuthenticate()==4){
+			return AjaxResult.error("暂无权限操作");
+		}
 		if (StringUtils.isNotNull(service.getSignInfoByLongitudeAndLatitude(signInfoPo))){
 			return AjaxResult.error("新增失败，该地点已有标牌");
 		}
@@ -97,6 +103,11 @@ public class SignController extends BaseController{
 	@ApiOperation(value = "删除地图标牌")
 	@DeleteMapping("/signInfo/{id}")
 	public AjaxResult deleteSignInfo(@PathVariable Long[] id){
+		LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
+		BaseUser baseUser=(BaseUser)loginUser.getUser();
+		if (baseUser.getAuthenticate()==4){
+			return AjaxResult.error("暂无权限操作");
+		}
 		return AjaxResult.success(service.deleteSignInfo(id));
 	}
